feat[design]: 디자인 시스템 적용
