+++ core/designsystem/build.gradle.kts
@@ -1,6 +1,7 @@
 @Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
 plugins {
     hmh("feature")
+    hmh("compose")
 }
 
 android {
